Implemented ViewPager Fragments
Now the main staggeredGridView has become part of a fragment activity.
Allows multiple tabs to be implemented later. Because of the use of
Fragment, the code has been changed a lot, staggeredGridView has been
moved out of MainActivity, and now inside the OnCreateView of a
Fragment.

New File structure  :
MainActivity -> PagerAdapter -> YiluMainListFragment & SecondFragment